# Conflicts:
#	include/battle.h
#	include/constants/battle_string_ids.h
#	include/pokemon.h
#	src/data/text/abilities.h
#	src/pokemon.c
#	src/pokemon_summary_screen.c

+++ src/pokemon.c
@@ -4802,7 +4802,7 @@ u8 GetMonsStateToDoubles_2(void)
     return (aliveCount > 1) ? PLAYER_HAS_TWO_USABLE_MONS : PLAYER_HAS_ONE_USABLE_MON;
 }
 
-u8 GetAbilityBySpecies(u16 species, u8 abilityNum, u8 formId)
+u16 GetAbilityBySpecies(u16 species, u8 abilityNum, u8 formId)
 {
     u16 formSpeciesId = GetFormSpeciesId(species, formId);
 
@@ -4816,7 +4816,7 @@ u8 GetAbilityBySpecies(u16 species, u8 abilityNum, u8 formId)
     return gLastUsedAbility;
 }
 
-u8 GetMonAbility(struct Pokemon *mon)
+u16 GetMonAbility(struct Pokemon *mon)
 {
     u16 species = GetMonData(mon, MON_DATA_SPECIES, NULL);
     u8 abilityNum = GetMonData(mon, MON_DATA_ABILITY_NUM, NULL);
